test: stabilize shared wait, review and Project lifecycle fixtures
(cherry picked from commit 850cce9eb5233df8e208006e136c17202470ea3f)
